requirements.txt
fastapi
uvicorn
langchain
pinecone-client
python-dotenv
